requirements.txt
joblib==1.3.2
numpy==1.25.2
opencv-python==4.8.0.76
pandas==2.1.0
Pillow==10.0.0
PyYAML==6.0.1
scikit-image==0.21.0
scikit-learn==1.3.0
cv2
xgboost
cupy-cuda12x
matplotlib